Move `i2c::Error` to dedicated module
I'm about to add another module for I2C slave mode which is going to
need access to `Error`. Leaving it were it is right now would introduce
a circular dependency.
